Updates in contracts & Initial Databases

+++ Geshotel/Geshotel.Web/Modules/Common/Migrations/DefaultDB/DefaultDB_20170123_163900_Initial.cs
@@ -33,7 +33,7 @@ namespace Geshotel.Migrations.DefaultDB
                 .ForeignKey("FK_cupos_tipo_habitacion_id", "tipos_habitacion", "tipo_habitacion_id")
                 .WithColumn("garantia").AsDecimal(5, 2).NotNullable().WithDefaultValue("0.00")
                 .WithColumn("reserva_automatica").AsBinary(1).NotNullable()
-                .WithColumn("user_id").AsInt16().Nullable()
+                .WithColumn("user_id").AsInt32().Nullable()
                 .WithColumn("fecha_modificacion").AsDateTime().Nullable();                
         }
         public override void Down()
